server/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
tensorflow-cpu==2.7.0
numpy==1.20.1
opencv-python-headless
gevent==21.12.0
bjoern==3.1.0
